Refactored the Config to be different from the CommandLineProcessor
